Finalized codes

+++ src/app/task-ui/task-ui.component.ts
@@ -40,8 +40,10 @@ export class TaskUiComponent implements OnInit {
 
 
   deleteTask(task_: Task) {
-    this.taskService.deleteTask(task_.id).subscribe(() => {
-      this.tasks = this.tasks.filter(task => !this.taskService.deletedtasks.includes(task.id));
-    });
+    if(confirm("Are you sure to delete the task "+ task_.title + " ?")) {
+      this.taskService.deleteTask(task_.id).subscribe(() => {
+        this.tasks = this.tasks.filter(task => !this.taskService.deletedtasks.includes(task.id));
+      });
+    }
   }
 }
